Migrate to latest biny mappings

+++ src/main/java/net/kozibrodka/tmt/TURBO_MODEL_164/TransformGroup.java
@@ -1,9 +1,9 @@
 package net.kozibrodka.tmt.TURBO_MODEL_164;
 
-import net.minecraft.util.maths.Vec3f;
+import net.minecraft.util.math.Vec3d;
 
 public abstract class TransformGroup {
     public abstract double getWeight();
 
-    public abstract Vec3f doTransformation(PositionTransformVertex var1);
+    public abstract Vec3d doTransformation(PositionTransformVertex var1);
 }
