+++ src/main/java/net/kozibrodka/tmt/TURBO_MODEL_173/TransformGroup.java
@@ -1,6 +1,6 @@
 package net.kozibrodka.tmt.TURBO_MODEL_173;
 
-import net.minecraft.util.maths.Vec3f;
+import net.minecraft.util.math.Vec3d;
 
 public abstract class TransformGroup
 {
@@ -11,6 +11,6 @@ public abstract class TransformGroup
 
     public abstract double getWeight();
 
-    public abstract Vec3f doTransformation(PositionTransformVertex positiontransformvertex);
+    public abstract Vec3d doTransformation(PositionTransformVertex positiontransformvertex);
 }
